Offload statement handling to a new Statement class
This separates the statements from the Table class, and allows multiple
query statements to be associated with a particular table for record
processing; future tables may need to prepare or otherwise manipulate
data requiring multiple queries.

This also removes the TDEF global hashref from the import code, replaced
by specific subroutines to add the query to the table, which is
effectively a proxy for Statement class instantiation.