Data.pm: comment-out VC table removal of matching sys_id rows
This code was likely added to allow for amendments that replaced the
vanity calls, but it breaks as the statement is invoked for EACH vanity
call, leaving just the final one when multiple were provided.

If this ends up causing a PK violation on commit when a future such
amendment is processed, smarter logic that tracks (perhaps in a temp
table) removed sys_id numbers, and scans the existing DB for a duplicate
the FIRST (and only first) time a new sys_id is found. This temp table
presumably would grow to the size of all unique sys_id values in the
input table source, as future instances are adding a call to the
application.